% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{factorDataSetDjia5Yrs}
\alias{factorDataSetDjia5Yrs}
\title{DJIA stocks Compustat factors 5yrs}
\source{
TBA
}
\usage{
data("factorDataSetDjia5Yrs")
}
\description{
Contains returns for 30 DJIA spanned across 9 Sectors - Energy, 
             ConsumerStaples, Industrials, Materials, Financials, InformationTechnology, HealthCare, ConsumerDiscretionary
             and Telecommunicationsstocks along with the 4 factors: MKTCAP, ENTVAL, P2B, EV2S, SIZE
             starting from Jan 2008 to Dec 2012.
}

