% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ffmTstats.R
\name{ffmTstats}
\alias{ffmTstats}
\title{t-stats and Plots for a fitted Fundamental Factor Model}
\usage{
ffmTstats(ffmObj, isPlot = TRUE, isPrint = FALSE, col = "cyan", lwd = 2,
  digits = 2, z.alpha = 1.96, layout = c(2, 3), type = "h", ...)
}
\arguments{
\item{ffmObj}{an object of class \code{ffm} produced by \code{fitFfm}}

\item{isPlot}{logical. If \code{FALSE} no plots are displayed.}

\item{isPrint}{logical. if \code{TRUE}, the time series of the computed factor model values is printed. Default is \code{FALSE},}

\item{col}{specification for the default plotting color. Default is cyan}

\item{lwd}{line width relative to the default. Default is 2.}

\item{digits}{an integer indicating the number of decimal places to be used for rounding. Default is 2.}

\item{z.alpha}{critical value corresponding to the confidence interval. Default is 1.96 i.e 95\% C.I}

\item{layout}{numeric vector of length 2 or 3 giving the number of columns, rows, and pages (optional) in the xyplot of t-statistics. Default is c(2,3).}

\item{type}{character. Type of the xyplot of t-statistics; \code{"l"} for lines, \code{"p"} for points, \code{"h"} for histogram like (or high-density) vertical lines
and \code{"b"} for both. Deafault is \code{"h"}.}

\item{...}{potentially further arguments passed.}
}
\value{
\code{ffmTstats} plots the t-stats and significant t-stats values  if \code{isPlot} is \code{TRUE} and returns a list with following components:
\item{tstats}{ an xts object of t-stats values.}
\item{z.alpha}{ critical value corresponding to the confidence interval.}
}
\description{
Calculate and plot the time series of the t-statistic values and the
 number of risk indices with significant t-stats for a fundamentally fit object.
}
\examples{
 
 data("factorDataSetDjia5Yrs")
 
#Fit a Ffm
 require(factorAnalytics)
 fit <- fitFfm(data = factorDataSetDjia5Yrs,exposure.vars = c("MKTCAP","ENTVAL","P2B","EV2S"),
               date.var = "DATE", ret.var = "RETURN", asset.var = "TICKER", fit.method="WLS")
               
 #Find tstats  with C.I = 95\% and plot the results.
 stats = ffmTstats(fit, isPlot = TRUE, col = "blue", digits =3, z.alpha =1.96)  
   
              

}
\author{
Doug Martin, Avinash Acharya
}

