% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/portRsq.R
\name{portRsqr}
\alias{portRsqr}
\title{R-squared and Adjusted R-squared for a Portfolio}
\usage{
portRsqr(ffmObj, rsq = T, rsqAdj = F, VIF = F, digits = 2,
  isPrint = F, ...)
}
\arguments{
\item{ffmObj}{an object of class \code{ffm} produced by \code{fitFfm}}

\item{rsq}{logical; if \code{TRUE}, R-squared values are computed for the portfolio. Default is \code{TRUE}.}

\item{rsqAdj}{logical; if \code{TRUE}, Adjusted R-squared values are computed for the portfolio. Default is \code{FALSE}.}

\item{VIF}{logical; if \code{TRUE}, Variance Inflation factor is calculated. Default is \code{FALSE}.
At least 2 continous variables are required to find VIF.}

\item{digits}{an integer indicating the number of decimal places to be used for rounding. Default is 2.}

\item{isPrint}{logical. if \code{TRUE}, the time series of the output is printed. Default is \code{FALSE},}

\item{...}{potentially further arguments passed.}
}
\value{
\code{portRsqr} returns the sample mean values and plots the time series of corresponding R squared values
                        and the Variance Inflation factors depending on the values of \code{rsq}, \code{rsqAdj} and \code{VIF}.
                        The time series of the output values are also printed if \code{isPrint} is \code{TRUE}
}
\description{
Calcluate and plot the R-squared, Adjusted R-squared and Variance Inflation Factors for a portfolio of assets
}
\examples{

#Load the data
 data("factorDataSetDjia5Yrs")

#Fit a Ffm
require(factorAnalytics)
 fit <- fitFfm(data=factorDataSetDjia5Yrs, asset.var="TICKER", ret.var="RETURN",
              date.var="DATE", exposure.vars="SECTOR")

#Calcuate and plot the portfolio R-squared values
 portRsqr(fit)
 
 fit1 <- fitFfm(data=factorDataSetDjia5Yrs, asset.var="TICKER", ret.var="RETURN",
              date.var="DATE", exposure.vars=c("SECTOR", "P2B", "EV2S", "MARKETCAP"))

#Plot and print the time series of VIF
 portRsqr(fit1, VIF=TRUE, isPrint=TRUE)
}
\author{
Avinash Acharya
}

