% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repRisk.R
\name{repRisk}
\alias{repRisk}
\alias{repRisk.ffm}
\alias{repRisk.tsfm}
\title{Decompose portfolio risk into individual factor contributions and provide tabular report}
\usage{
repRisk(object, ...)

\method{repRisk}{tsfm}(object, weights = NULL, risk.factor = c("Sd", "VaR",
  "Es"), risk.budget = c("RM", "FMCR", "FCR", "FPCR"), nrowPrint = 20,
  p = 0.95, type = c("np", "normal"), use = "pairwise.complete.obs", ...)

\method{repRisk}{ffm}(object, weights = NULL, risk.factor = c("Sd", "VaR",
  "Es"), risk.budget = c("RM", "mRM", "cRM", "pRM"), nrowPrint = 20,
  p = 0.95, type = c("np", "normal"), ...)
}
\arguments{
\item{object}{fit object of class \code{tsfm}, or \code{ffm}.}

\item{...}{optional arguments passed to \code{\link[stats]{cov}}.}

\item{weights}{a vector of weights of the assets in the portfolio. Default is NULL, 
in which case an equal weights will be used.}

\item{use}{an optional character string giving a method for computing 
covariances in the presence of missing values. This must be (an 
abbreviation of) one of the strings "everything", "all.obs", 
"complete.obs", "na.or.complete", or "pairwise.complete.obs". Default is 
"pairwise.complete.obs".}
}
\value{
A table containing 
\item{risk.budget = 'RM'}{length-(N + 1) vector of factor model risk measure of portfolio return 
as well assets return.}
\item{risk.budget = 'FMCR'}{(N + 1) * (K + 1) matrix of marginal contributions to risk of portfolio 
return as well assets return.}
\item{risk.budget = 'FCR'}{(N + 1) * (K + 1) matrix of component contributions to risk of portfolio 
return as well assets return.}
\item{risk.budget = 'FPCR'}{(N + 1) * (K + 1) matrix of percentage component contributions to risk 
of portfolio return as well assets return.}
Where, K is the number of factors, N is the number of assets.
}
\description{
Compute the factor contributions to standard deviation (SD), Value-at-Risk (VaR), 
Expected Tail Loss or Expected Shortfall (ES) of the return of individual asset within a portfolio 
return of a portfolio based on Euler's theorem, given the fitted factor model.
}
\examples{
# Time Series Factor Model
data(managers)
fit.macro <- factorAnalytics::fitTsfm(asset.names=colnames(managers[,(1:6)]),
                     factor.names=colnames(managers[,(7:9)]),
                     rf.name="US.3m.TR", data=managers)
report <- repRisk(fit.macro, risk.factor = "Es", risk.budget = 'FPCR', nrowPrint = 10)
report 

# random weights 
wts = runif(6)
wts = wts/sum(wts)
names(wts) <- colnames(managers)[1:6]
repRisk(fit.macro, risk.factor = "VaR", risk.budget = 'FMCR', nrowPrint = 10)

# Fundamental Factor Model
data("stocks145scores6")
dat = stocks145scores6
dat$DATE = as.yearmon(dat$DATE)
dat = dat[dat$DATE >=as.yearmon("2008-01-01") & dat$DATE <= as.yearmon("2012-12-31"),]

# Load long-only GMV weights for the return data
data("wtsStocks145GmvLo")
wtsStocks145GmvLo = round(wtsStocks145GmvLo,5)  
                                                     
# fit a fundamental factor model
fit.cross <- fitFfm(data = dat, 
              exposure.vars = c("SECTOR","ROE","BP","PM12M1M","SIZE","ANNVOL1M","EP"),
              date.var = "DATE", ret.var = "RETURN", asset.var = "TICKER", 
              fit.method="WLS", z.score = TRUE)
repRisk(fit.cross, risk.factor = "Sd", risk.budget = 'FCR', nrowPrint = 10) 
# get the factor contributions of risk 
decomp$cSd
repRisk(fit.cross, wtsStocks145GmvLo, risk.factor = "Sd", risk.budget = 'FPCR', nrowPrint = 10)               
 
}
\author{
Douglas Martin, Lingjie Yi
}
\seealso{
\code{\link{fitTsfm}}, \code{\link{fitSfm}}, \code{\link{fitFfm}}
for the different factor model fitting functions.

\code{\link{portVaRDecomp}} for factor model VaR decomposition.
\code{\link{portEsDecomp}} for factor model ES decomposition.
}

