% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{stocks145scores6}
\alias{stocks145scores6}
\title{CRSP stocks Capital IQ scores}
\source{
TBA
}
\usage{
data("stocks145scores6")
}
\description{
Contains returns for 145 stocks starting from Jan 1990 to Dec 2014 spanned across 10 Sectors- Energy, 
             ConsumerStaples, Industrials, Materials, Financials, InformationTechnology, HealthCare, ConsumerDiscretionary,
             Utilities and Telecommunications along with 6 factors: ROE, BP, MOM121, SIZE, VOL121, EP
}

